// src/types/openapi_MS.ts
export interface components {
    schemas: {
// ...
        ResGradeDTO: {
            /** Format: int64 */
            id?: number;
            name?: string;
            /** Format: int64 */
            studentsInPeriodCount?: number;
        };
// ...
        ResStudyWeekDTO: {
            /** Format: uuid */
            week_uuid?: string;
            /** Format: int32 */
            week_number?: number;
            /** Format: int32 */
            school_year?: number;
            /** Format: date */
            week_start_date?: string;
            /** Format: date */
            week_end_date?: string;
            /** Format: date-time */
            created_at?: string;
            /** Format: date-time */
            updated_at?: string;
            created_by?: string;
            updated_by?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}